Minor code reorganization and add some unit tests.
