+++ core/js/src/index.ts
@@ -3,11 +3,13 @@ export * from "./http_headers";
 export * from "./merge_row_batch";
 export * from "./object";
 export * from "./score";
-export * from "./util";
+export * from "./json_util";
+export * from "./object_util";
+export * from "./string_util";
+export * from "./type_util";
 export * from "./span_identifier_v1";
 export * from "./span_identifier_v2";
 export * from "./span_types";
 export * from "./git_fields";
 export * from "./xact-ids";
 export * from "./zod_util";
-export * from "./json_util";
