+++ core/js/src/json_util.ts
@@ -2,7 +2,7 @@
 // serialized in a deterministic key order.
 //
 // https://gist.github.com/davidfurlong/463a83a33b70a3b6618e97ec9679e490
-export function sortedKeyReplacer(_key: string, value: any) {
+export function deterministicReplacer(_key: string, value: any) {
   return value instanceof Object && !(value instanceof Array)
     ? Object.keys(value)
         .sort()
@@ -11,4 +11,8 @@ export function sortedKeyReplacer(_key: string, value: any) {
           return sorted;
         }, {})
     : value;
+}
+
+export function constructJsonArray(items: string[]) {
+  return `[${items.join(",")}]`;
 }
