+++ core/js/src/object_util.ts
@@ -1,3 +1,5 @@
+import { isObject } from "./type_util";
+
 // Mutably updates `mergeInto` with the contents of `mergeFrom`, merging objects
 // deeply. Does not merge any further than `merge_paths`.
 export function mergeDictsWithPaths({
@@ -101,37 +103,6 @@ export function forEachMissingKey({
   helper(lhs, rhs, []);
 }
 
-export function capitalize(s: string, sep?: string) {
-  const items = sep ? s.split(sep) : [s];
-  return items
-    .map((s) => (s ? s.charAt(0).toUpperCase() + s.slice(1) : s))
-    .join(sep || "");
-}
-
-export function snakeToCamelCase(s: string) {
-  return s
-    .split("_")
-    .map((s) => capitalize(s))
-    .join("");
-}
-
-export function snakeToTitleCase(s: string) {
-  return capitalize(s, "_").replace("_", " ");
-}
-
-export function constructJsonArray(items: string[]) {
-  return `[${items.join(",")}]`;
-}
-
-export function _urljoin(...parts: string[]): string {
-  return parts
-    .map((x, i) =>
-      x.replace(/^\//, "").replace(i < parts.length - 1 ? /\/$/ : "", ""),
-    )
-    .filter((x) => x.trim() !== "")
-    .join("/");
-}
-
 export function mapAt<K, V>(m: Map<K, V>, k: K): V {
   const ret = m.get(k);
   if (ret === undefined) {
@@ -187,26 +158,6 @@ export function recordSetDefault<K extends string | number | symbol, V>(
   } else {
     return ret;
   }
-}
-
-/**
- * @example
- * isObject([1, 2, 3]) // false
- */
-export function isObject(value: any): value is { [key: string]: any } {
-  return value instanceof Object && !(value instanceof Array);
-}
-
-export function isArray(value: any): value is any[] {
-  return value instanceof Array;
-}
-
-export function isEmpty(a: any): a is null | undefined {
-  return a === undefined || a === null;
-}
-
-export function isNumber(a: any) {
-  return typeof a === "number" || typeof a === "bigint";
 }
 
 /**
